adds concurrency test and updates code to address possible issues (#9)
Need more synchronization in calls to tunnel streams since Go docs
for grpc streams explicitly states that concurrent calls to Send
and concurrent calls to Recv are not safe. (But a call to Send that
is concurrent with a call to Recv *is* safe.)

Also applies a fix that was found and patched in jh/flow-control branch
related to erroneous non-sequential creation of stream IDs.

Unrelated: removed the predeclared checker because it has an
internal error and appears to no longer be maintained.